% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqplot.R
\name{freqplot}
\alias{freqplot}
\title{A plotting function for SISPA sample identifiers}
\usage{
freqplot(x)
}
\arguments{
\item{x}{A data frame containing samples as rows followed by zscores and estimated changepoints to be plotted.}
}
\value{
Bar plot pdf illustrating distribution of samples
}
\description{
Given a sample changepoint data frame, will plot number of samples with and without profile activity
}
\details{
This function expects the output from cptSamples function of SISPA package, and shows the number of samples with (orange filled bars) and without profile activity (grey filled bars).
}
\examples{
samples <- c("s1","s2","s3","s4","s5","s6","s7","s8","s9","s10")
zscores <- c(3.83,2.70,2.67,2.31,1.70,1.25,-0.42,-1.01,-2.43,-3.37)
changepoints <- c(1,1,1,2,2,3,3,NA,NA,NA)
sample_groups <- c(1,1,1,0,0,0,0,0,0,0)
my.data = data.frame(samples,zscores,changepoints,sample_groups)
freqplot(my.data)
}

